+++ Arduino/lib/PID_Controller/src/PID_Controller.h
@@ -6,24 +6,24 @@
 class pid_controller {
   private:
     int _T, _r, _u;
-    double _Kp, _Ki, _Kaw;
-    double _y, _e, _p, _i, _aw;
-    double _y_prev, _e_prev, _i_prev;
+    float _Kp, _Ki, _Kaw;
+    float _y, _e, _p, _i, _aw;
+    float _y_prev, _e_prev, _i_prev;
 
     void _deadzone();
   public:
     // Constructor
-    pid_controller(int T, int r, double Kp, double Ki, double Kaw);
+    pid_controller(int T, int r, float Kp, float Ki, float Kaw);
 
-    double process();
-    int saturate(double aw);
+    float process();
+    int saturate(float aw);
     void flush();
 
     int getRef();
     void setRef(int r);
-    double getErr();
-    double getLight();
-    void setLight(double y);
+    float getErr();
+    float getLight();
+    void setLight(float y);
 };
 
 #endif
